# swagger.yaml
...
swagger: "2.0"
info:
  description: "The inner API of Popcube aim to define every methods need to manage\
    \ communication organisation inside PopCube chat project"
  version: "alpha"
  title: "Popcube iner API"
  contact:
    name: "Titouan Freville"
    email: "titouanfreville@gmail.com"
host: "api.popcube"
basePath: "/alpha"
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /avatar:
    get:
      tags:
      - "Avatars"
      summary: "Get avatars"
      description: "This will get all the avatars available in the organisation."
      operationId: "getAllAvatar"
      parameters: []
      responses:
        200:
          description: "avatarSlice Array of avatars"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Avatar"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
    post:
      tags:
      - "Avatars"
      summary: "New avatar"
      description: "This will create an avatar for organisation avatars library."
      operationId: "newAvatar"
      parameters:
      - in: "body"
        name: "avatar"
        description: "Link of the avatar in server."
        required: false
        schema:
          $ref: "#/definitions/Avatar"
        x-go-name: "Avatar"
      responses:
        200:
          description: "avatarObjectSuccess list of avatars"
          schema:
            $ref: "#/definitions/Avatar"
        422:
          description: "wrongEntityError is an error object to inform that the provided\
            \ object was not correctly formated"
          schema:
            $ref: "#/definitions/genericErrorModel"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /avatar/all:
    get:
      tags:
      - "Avatars"
      summary: "Get avatars"
      description: "This will get all the avatars available in the organisation."
      operationId: "getAllAvatar1"
      parameters: []
      responses:
        200:
          description: "avatarSlice Array of avatars"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Avatar"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /avatar/link/{avatarLink}:
    get:
      tags:
      - "Avatars"
      summary: "Get avatar from link"
      description: "This will return the avatar object corresponding to provided link"
      operationId: "getAvatarFromLink"
      parameters:
      - name: "avatarLink"
        in: "path"
        description: "Link of the avatar in server."
        required: true
        type: "string"
        x-go-name: "AvatarLink"
      responses:
        200:
          description: "avatarObjectSuccess list of avatars"
          schema:
            $ref: "#/definitions/Avatar"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /avatar/name/{avatarName}:
    get:
      tags:
      - "Avatars"
      summary: "y\nGet avatar from name"
      description: "This will return the avatar object corresponding to provided name"
      operationId: "getAvatarFromName"
      parameters:
      - name: "avatarName"
        in: "path"
        description: "Link of the avatar in server."
        required: true
        type: "string"
        x-go-name: "AvatarName"
      responses:
        200:
          description: "avatarObjectSuccess list of avatars"
          schema:
            $ref: "#/definitions/Avatar"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /avatar/new:
    post:
      tags:
      - "Avatars"
      summary: "New avatar"
      description: "This will create an avatar for organisation avatars library."
      operationId: "newAvatar1"
      parameters:
      - in: "body"
        name: "avatar"
        description: "Link of the avatar in server."
        required: false
        schema:
          $ref: "#/definitions/Avatar"
        x-go-name: "Avatar"
      responses:
        200:
          description: "avatarObjectSuccess list of avatars"
          schema:
            $ref: "#/definitions/Avatar"
        422:
          description: "wrongEntityError is an error object to inform that the provided\
            \ object was not correctly formated"
          schema:
            $ref: "#/definitions/genericErrorModel"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /avatar/{avatarID}:
    put:
      tags:
      - "Avatars"
      summary: "Update avatar"
      description: "This will return the new avatar object"
      operationId: "updateAvatar"
      parameters:
      - name: "avatarID"
        in: "path"
        description: "Link of the avatar in server."
        required: true
        type: "integer"
        format: "int64"
        x-go-name: "AvatarID"
      - in: "body"
        name: "avatar"
        description: "Link of the avatar in server."
        required: false
        schema:
          $ref: "#/definitions/Avatar"
        x-go-name: "Avatar"
      responses:
        200:
          description: "avatarObjectSuccess list of avatars"
          schema:
            $ref: "#/definitions/Avatar"
        422:
          description: "wrongEntityError is an error object to inform that the provided\
            \ object was not correctly formated"
          schema:
            $ref: "#/definitions/genericErrorModel"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
...
definitions:
  Avatar:
    type: "object"
    required:
    - "link"
    - "name"
    properties:
      id:
        type: "integer"
        format: "uint64"
        description: "id of the avatar"
        minimum: 0.0
        x-go-name: "IDAvatar"
      link:
        type: "string"
        description: "Path into server"
        x-go-name: "Link"
      name:
        type: "string"
        description: "Avatar name"
        x-go-name: "Name"
    title: "Avatar object"
    description: "Avatar object store default avatar images you can use within an\
      \ association.\nRequired apply only for creation of the object. Id is required\
      \ only on update."
    x-go-package: "github.com/titouanfreville/popcubeapi/models"
...
  genericErrorModel:
    type: "object"
    properties:
      detailed_error:
        type: "string"
        description: "Detail of error"
        x-go-name: "DetailedError"
      id:
        type: "string"
        description: "ID for the error"
        x-go-name: "ID"
      is_oauth:
        type: "boolean"
        x-go-name: "IsOAuth"
      message:
        type: "string"
        description: "Error's message"
        x-go-name: "Message"
      request_id:
        type: "string"
        description: "Id of the request if exist"
        x-go-name: "RequestID"
      status_code:
        type: "integer"
        format: "int64"
        description: "EgenericErrorModelrror's status code"
        x-go-name: "StatusCode"
    title: "AppError Type used to structure error reporting for popcube chat project."
    x-go-name: "AppError"
    x-go-package: "github.com/titouanfreville/popcubeapi/utils"
...